refactor: routes into own dir
